+++ lib/chatview.dart
@@ -22,12 +22,6 @@
 
 library chatview;
 
-export 'src/widgets/chat_view.dart';
-export 'src/models/models.dart';
-export 'src/widgets/chat_view_appbar.dart';
-export 'src/values/enumeration.dart';
-export 'src/values/typedefs.dart';
-export 'package:chatview_utils/chatview_utils.dart';
 export 'package:audio_waveforms/audio_waveforms.dart'
     show
         WaveStyle,
@@ -35,6 +29,17 @@ export 'package:audio_waveforms/audio_waveforms.dart'
         AndroidEncoder,
         IosEncoder,
         AndroidOutputFormat;
-export 'src/models/config_models/receipts_widget_config.dart';
-export 'src/extensions/extensions.dart' show MessageTypes;
+export 'package:chatview_utils/chatview_utils.dart';
 export 'package:emoji_picker_flutter/emoji_picker_flutter.dart';
+
+export 'src/extensions/extensions.dart' show MessageTypes;
+export 'src/models/config_models/receipts_widget_config.dart';
+export 'src/models/models.dart';
+export 'src/values/enumeration.dart';
+export 'src/values/typedefs.dart';
+export 'src/widgets/action_widgets/camera_action_button.dart';
+export 'src/widgets/action_widgets/gallery_action_button.dart';
+export 'src/widgets/action_widgets/overlay_action_button.dart';
+export 'src/widgets/action_widgets/text_field_action_button.dart';
+export 'src/widgets/chat_view.dart';
+export 'src/widgets/chat_view_appbar.dart';